requirements.txt
python-dotenv
openai
langchain
whisper
numpy
more-itertools
transformers
ffmpeg-python
pyaudio
SpeechRecognition
pydub
pyttsx3